ai-agent/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langgraph==0.0.69
langchain==0.2.6
langchain-openai==0.1.9
langchain-community==0.2.6
langchain-core==0.2.10
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.4
